Move distributed-manager
